+++ src/shared/types/error.codes.js
@@ -5,6 +5,7 @@ export const ERROR_CODES = {
   FORBIDDEN: "FORBIDDEN",
   UPLOAD_ERROR: "UPLOAD_ERROR",
   PAGINATION_VALIDATION: "PAGINATION_VALIDATION_ERROR",
+  ROUTE_SERVICE_UNAVAILABLE: "ROUTE_SERVICE_UNAVAILABLE",
 
   COMPLAINT_NOT_FOUND: "COMPLAINT_NOT_FOUND",
   COMPLAINT_VALIDATION: "COMPLAINT_VALIDATION_ERROR",
